requirements.txt
langchain
langchain-core
langchain-openai
langgraph
pydantic
python-dotenv
requests
beautifulsoup4
python-docx
pypdf
streamlit
rank-bm25
faiss-cpu
numpy
python-jobspy
